+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Angular 20 support
 
+### Changed
+- Migrate ESLint config v8 to v9
+
 ## [19.0.0] - 2025-07-11
 ### Added
 - Angular 19 support